[FIX] : Backend Fixes done

+++ backend/sql/create/create_users_table.sql
@@ -4,7 +4,7 @@ CREATE TABLE
         first_name TEXT NOT NULL,
         last_name TEXT NOT NULL,
         email TEXT NOT NULL,
-        phone INTEGER NOT NULL,
+        phone BIGINT NOT NULL,
         created_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
         updated_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP
     )
